churchdir/requirements.txt
Django==5.0.6
django-import-export==4.1.1
gunicorn==21.2.0
